[layouts/type-stage] rename from "TypeSpec -- Ramp" or just /layouts/ramp to layouts/type-stage
