+++ lib/js/components/layouts/type-stage/defaults.mjs
