+++ lib/js/components/layouts/type-stage/properties-generators.mjs
