+++ lib/js/components/layouts/type-stage/shared.typeroof.jsx
