+++ lib/js/components/layouts/type-stage/synthetic-values.mjs
