+++ lib/js/components/layouts/type-stage/type-specnion.mjs
