+++ lib/js/components/layouts/type-tools-grid.mjs
@@ -140,18 +140,18 @@ import {
 
 import {
     LocalScopeTypeSpecnion
-} from './ramp/type-specnion.mjs';
+} from './type-stage/type-specnion.mjs';
 import {
     axisLocationsGen
   , openTypeFeaturesGen
   , languageTagGen
-} from './ramp/properties-generators.mjs';
+} from './type-stage/properties-generators.mjs';
 import {
     TypeSpecLiveProperties
-} from './ramp/live-properties.typeroof.jsx';
+} from './type-stage/live-properties.typeroof.jsx';
 import {
     SyntheticValue
-} from './ramp/synthetic-values.mjs';
+} from './type-stage/synthetic-values.mjs';
 
 import {
 //    getPropertyValue
