+++ lib/js/components/main-ui.mjs
@@ -33,7 +33,7 @@ import * as PentrekLayout from './dev-layouts/pentrek.mjs';
 import * as StageAndActorsLayout from './layouts/stage-and-actors.mjs';
 import * as VideoproofArrayLayout from './dev-layouts/videoproof-array.mjs';
 import * as VideoproofLayout from './layouts/videoproof.typeroof.jsx';
-import * as TypeSpecRamp from './layouts/ramp/index.typeroof.jsx';
+import * as TypeStage from './layouts/type-stage/index.typeroof.jsx';
 import * as SimpleRamp from './layouts/simple-ramp/index.typeroof.jsx';
 import * as TypeToolsGrid from './layouts/type-tools-grid.mjs';
 import * as ExampleReactLayout from './dev-layouts/example-react.mjs';
@@ -67,7 +67,7 @@ export const Layouts = Object.freeze([
   , ['VideoproofArrayDev', 'Videoproof Array (Interim)', VideoproofArrayLayout, LAYOUT_GROUP_DEVELOPMENT]
   , ['Videoproof', 'Video Proof', VideoproofLayout, LAYOUT_GROUP_VIDEO_PROOF]
   , ['TypeToolsGrid', 'Grid', TypeToolsGrid, LAYOUT_GROUP_TYPE_TOOLS]
-  , ['TypeSpecRamp', 'TypeSpec – Ramp', TypeSpecRamp, LAYOUT_GROUP_TYPE_TOOLS]
+  , ['TypeStage', 'Type Stage', TypeStage, LAYOUT_GROUP_DEFAULT]
   , ['SimpleRamp', 'Simple Ramp', SimpleRamp, LAYOUT_GROUP_TYPE_TOOLS]
 ]);
