+++ lib/js/main-player.mjs
@@ -34,7 +34,7 @@ import {
 import {PlayerLayout as StageAndActorsLayout} from './components/layouts/stage-and-actors.mjs';
 import * as VideoproofArrayLayout from './components/dev-layouts/videoproof-array.mjs';
 import * as VideoproofLayout from './components/layouts/videoproof.typeroof.jsx';
-import * as TypeSpecRamp from './components/layouts/ramp/index.typeroof.jsx';
+import * as TypeStage from './components/layouts/type-stage/index.typeroof.jsx';
 import * as TypeToolsGrid from './components/layouts/type-tools-grid.mjs';
 
 const LAYOUT_GROUP_DEFAULT = Symbol('LAYOUT_GROUP_DEFAULT')
@@ -55,7 +55,7 @@ export const Layouts = Object.freeze([
   , ['VideoproofArrayDev', 'Videoproof Array (Interim)', VideoproofArrayLayout, LAYOUT_GROUP_DEFAULT]
   , ['Videoproof', 'Video Proof', VideoproofLayout, LAYOUT_GROUP_DEFAULT]
   , ['TypeToolsGrid', 'Grid', TypeToolsGrid, LAYOUT_GROUP_DEFAULT]
-  , ['TypeSpecRamp', 'TypeSpec - Ramp', TypeSpecRamp, LAYOUT_GROUP_DEFAULT]
+  , ['TypeStage', 'Type Stage', TypeStage, LAYOUT_GROUP_DEFAULT]
 ]);
 
 function _addListener(node, eventName, callback, ...args) {
